Fix issues

+++ kernel/syscalls/000_read.c
@@ -84,7 +84,7 @@ long sys_read (unsigned int fd, void __user * buf, size_t size) {
 
 
 
-    int e = 0;
+    ssize_t e = 0;
 
     __lock(&current_task->fd->descriptors[fd].ref->lock, {
